feat: move errorMap
