+++ packages/astro/test/units/content-collections/error-map.test.js
@@ -1,6 +1,6 @@
 import * as assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import { errorMap } from '../../../dist/content/index.js';
+import { errorMap } from '../../../dist/core/errors/index.js';
 import { z } from '../../../zod.mjs';
 import { fixLineEndings } from '../../test-utils.js';
